backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
mysql-connector-python==8.3.0
requests==2.31.0
httpx==0.27.0
python-dotenv==1.0.0
openai==1.0.0
supabase
pytest==7.4.2

# Multi-tenancy and Security
psycopg2-binary==2.9.9      # PostgreSQL adapter
python-jose[cryptography]==3.3.0  # JWT handling
passlib[bcrypt]==1.7.4      # Password hashing
python-multipart==0.0.6     # Form data parsing
bcrypt==3.2.2               # Password hashing backend (compatible with passlib 1.7.4)
email-validator             # Email validation for Pydantic
